requirements.txt
asgiref==3.2.7
astroid==2.4.1
beautifulsoup4==4.9.1
Django==3.0.7
django-bootstrap4==2.0.1
django-tinymce==3.0.2
factory-boy==2.12.0
Faker==4.1.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
psycopg2==2.8.5
pylint==2.5.2
pylint-django==2.0.15
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
wrapt==1.12.1
